bce-iot 10936: support multifile synchronization (#445)

+++ baetyl-agent/download.go
@@ -16,7 +16,7 @@ func (a *agent) downloadVolumes(volumes []baetyl.VolumeInfo) error {
 		if v.Meta.URL == "" {
 			continue
 		}
-		_, _, err := a.downloadVolume(v)
+		_, _, err := a.downloadVolume(v, "", true)
 		if err != nil {
 			return err
 		}
@@ -24,19 +24,21 @@ func (a *agent) downloadVolumes(volumes []baetyl.VolumeInfo) error {
 	return nil
 }
 
-func (a *agent) downloadVolume(v baetyl.VolumeInfo) (string, string, error) {
+func (a *agent) downloadVolume(v baetyl.VolumeInfo, name string, zip bool) (string, string, error) {
 	rp, err := filepath.Rel(baetyl.DefaultDBDir, v.Path)
 	if err != nil {
 		return "", "", fmt.Errorf("path of volume (%s) invalid: %s", v.Name, err.Error())
 	}
-
 	hostDir := path.Join(baetyl.DefaultDBDir, rp)
 	containerDir := path.Join(baetyl.DefaultDBDir, "volumes", rp)
-	containerZipFile := path.Join(containerDir, v.Name+".zip")
+	if name == "" {
+		name = v.Name + ".zip"
+	}
+	containerFile := path.Join(containerDir, name)
 
 	// volume exists
-	if utils.FileExists(containerZipFile) {
-		md5, err := utils.CalculateFileMD5(containerZipFile)
+	if utils.FileExists(containerFile) {
+		md5, err := utils.CalculateFileMD5(containerFile)
 		if err == nil && md5 == v.Meta.MD5 {
 			a.ctx.Log().Debugf("volume (%s) exists", v.Name)
 			return hostDir, containerDir, nil
@@ -58,13 +60,13 @@ func (a *agent) downloadVolume(v baetyl.VolumeInfo) (string, string, error) {
 	if err != nil {
 		return "", "", fmt.Errorf("failed to prepare volume (%s): %s", v.Name, err.Error())
 	}
-	err = utils.WriteFile(containerZipFile, res)
+	err = utils.WriteFile(containerFile, res)
 	if err != nil {
 		os.RemoveAll(containerDir)
 		return "", "", fmt.Errorf("failed to prepare volume (%s): %s", v.Name, err.Error())
 	}
 
-	md5, err := utils.CalculateFileMD5(containerZipFile)
+	md5, err := utils.CalculateFileMD5(containerFile)
 	if err != nil {
 		os.RemoveAll(containerDir)
 		return "", "", fmt.Errorf("failed to calculate MD5 of volume (%s): %s", v.Name, err.Error())
@@ -74,10 +76,12 @@ func (a *agent) downloadVolume(v baetyl.VolumeInfo) (string, string, error) {
 		return "", "", fmt.Errorf("MD5 of volume (%s) invalid", v.Name)
 	}
 
-	err = utils.Unzip(containerZipFile, containerDir)
-	if err != nil {
-		os.RemoveAll(containerDir)
-		return "", "", fmt.Errorf("failed to unzip volume (%s): %s", v.Name, err.Error())
+	if zip {
+		err = utils.Unzip(containerFile, containerDir)
+		if err != nil {
+			os.RemoveAll(containerDir)
+			return "", "", fmt.Errorf("failed to unzip volume (%s): %s", v.Name, err.Error())
+		}
 	}
 	return hostDir, containerDir, nil
 }
